requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.14
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
cachetools==5.5.0
cffi==1.17.1
cfgv==3.4.0
clang-format==17.0.6
clean-fid==0.1.35
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cpplint==1.6.1
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
datasets==3.1.0
decorator==4.4.2
deepspeed==0.14.4
descartes==1.1.0
diffusers==0.31.0
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
einops==0.8.0
eval_type_backport==0.2.0
evaluate==0.4.3
facexlib==0.3.0
fairscale==0.4.13
faiss==1.8.0
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.14.0
filterpy==1.4.5
fire==0.7.0
flake8==3.8.2
flake8-bugbear==23.3.12
flake8-comprehensions==3.16.0
flake8-executable==2.1.3
flake8-pyi==20.5.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gradio==5.0.1
gradio_client==1.4.0
graphviz==0.20.3
grpcio==1.65.4
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
icecream==2.1.3
identify==2.6.2
imageio==2.35.0
imageio-ffmpeg==0.5.1
imgaug==0.4.0
iniconfig==2.0.0
jmespath==0.10.0
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.5
langid==1.1.6
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.6
llvmlite==0.43.0
lmdb==1.5.1
lpips==0.1.4
Markdown==3.6
markdown-it-py==3.0.0
matplotlib==3.5.3
mccabe==0.6.1
mdurl==0.1.2
mkl-service==2.4.0
more-itertools==10.5.0
moviepy==1.0.3
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
MultiScaleDeformableAttention==1.0
ninja==1.11.1.1
nltk==3.9.1
nodeenv==1.9.1
numba==0.60.0
numpy
nuscenes-devkit==1.1.11
nvidia-ml-py==12.560.30
omegaconf==2.3.0
openai-clip==1.0.1
openai-whisper==20231117
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openxlab==0.1.2
orjson==3.10.7
oss2==2.17.0
pandas==2.2.2
patsy==0.5.6
peft==0.12.0
pluggy==1.5.0
plyfile==1.1
pooch==1.8.2
pre-commit==3.5.0
proglog==0.1.10
protobuf==5.27.3
pudb==2024.1.2
py-cpuinfo==9.0.0
pyarrow==18.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycodestyle==2.6.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyflakes==2.2.0
pyiqa==0.1.12
pyparsing==3.1.2
pyquaternion==0.9.9
pytest==8.3.3
python-multipart==0.0.12
pytorch-fid==0.3.0
pytorch-lightning==2.4.0
pytz==2023.4
regex==2024.7.24
requests==2.32.3
rich==13.4.2
rouge-score==0.1.2
ruff==0.6.9
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.13.0
setproctitle==1.3.3
Shapely==1.8.5.post1
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.38.6
statsmodels==0.14.4
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
TextGrid==1.6.1
threadpoolctl==3.5.0
tifffile==2024.8.10
tiktoken==0.8.0
timm==1.0.8
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.0
torchaudio==2.4.0
torchmetrics==1.4.1
torchsummary==1.5.1
torchvision==0.19.0
torchviz==0.0.3
tqdm==4.67.0
transformers==4.44.2
triton==2.3.1
typer==0.12.5
tzdata==2024.1
urllib3==1.26.20
urwid==2.6.15
urwid_readline==0.14
uvicorn==0.31.1
virtualenv==20.27.1
websockets==12.0
Werkzeug==3.0.3
xformers==0.0.27.post2
# Editable install with no version control (xllmx==0.0.1)
xxhash==3.5.0
yapf==0.40.2
yarl==1.9.4
